u32 -> i32

+++ src/tokenizer.rs
@@ -4,7 +4,7 @@ use std::{iter::Peekable, str::Chars};
 pub enum Token {
     // Primitive
     Identifier(String),
-    Integer(u32),
+    Integer(i32),
     String(String),
 
     // Keywords
@@ -165,13 +165,13 @@ impl<'a> Tokenizer<'a> {
     }
 
     fn read_integer(&mut self, nextc: char) -> Token {
-        let mut value: u32 = (nextc as u32) - ('0' as u32);
+        let mut value: i32 = (nextc as i32) - ('0' as i32);
         self.iter.next();
 
         loop {
             match self.peek_char() {
                 Some(x @ '0'..='9') => {
-                    let n = (*x as u32) - ('0' as u32);
+                    let n = (*x as i32) - ('0' as i32);
 
                     value = value * 10 + n;
                 }
